Forker changes for Cray Port

+++ src/lib/Components/system_script_forker.py
@@ -1,51 +1,67 @@
+"""System script forker classes.  This forker is intended for scripts that are
+run internally to cobalt.  This forker runs these scripts as root and does not
+setuid/setgid down to another user.
+
+
+"""
+
 import logging
 import os
 import tempfile
 
 import Cobalt
+import Cobalt.Util
+
 import Cobalt.Components.base_forker
 BaseForker = Cobalt.Components.base_forker.BaseForker
 BaseChild = Cobalt.Components.base_forker.BaseChild
-import Cobalt.Util
 convert_argv_to_quoted_command_string = Cobalt.Util.convert_argv_to_quoted_command_string
 
 _logger = logging.getLogger(__name__.split('.')[-1])
 
 
 class SystemScriptChild (BaseChild):
+    '''System script forker children.  This extends the base child to take
+    strings to stdin and return strings to stdout of a forked process.  A file
+    may be used for stdin/stdout buffering as well.  Use the string-based
+    methods to avoid filesystem operations for rapid child script requests.
+
+    '''
+
     def __init__(self, id = None, **kwargs):
         BaseChild.__init__(self, id=id, **kwargs)
-
-        try:
-            self.stdin_file = open("/dev/null")
-        except (OSError, IOError), e:
-            _logger.error("%s: unable to open /dev/null (to redirect stdin): %s", self.label, e)
-            raise
-
-        stdout_fn = None
-        try:
+        if self.stdin_string is None:
             try:
-                stdout_fd, stdout_fn = tempfile.mkstemp(prefix="cobalt_ssf_%s_" % (self.id,), suffix=".stdout")
+                self.stdin_file = open("/dev/null")
             except (OSError, IOError), e:
-                _logger.error("%s: unable to create temporary stdout file: %s", self.label, e)
-            else:
-                try:
-                    self.stdout_file = os.fdopen(stdout_fd, 'a+', 1)
-                except (OSError, IOError), e:
-                    _logger.error("%s: unable to open temporary stdout file: %s", self.label, e)
-        finally:
-            if stdout_fn is not None:
-                try:
-                    os.unlink(stdout_fn)
-                except (OSError, IOError), e:
-                    _logger.warning("%s: unable to remove temporary stdout file: %s", self.label, e)
-        if self.stdout_file is None:
-            try:
-                _logger.warning("%s: redirecting stdout to /dev/null", self.label)
-                self.stdout_file = open("/dev/null")
-            except (OSError, IOError), e:
-                _logger.error("%s: unable to open /dev/null (to redirect stdout): %s", self.label, e)
+                _logger.error("%s: unable to open /dev/null (to redirect stdin): %s", self.label, e)
                 raise
+
+        if not self.use_stdout_string:
+            stdout_fn = None
+            try:
+                try:
+                    stdout_fd, stdout_fn = tempfile.mkstemp(prefix="cobalt_ssf_%s_" % (self.id,), suffix=".stdout")
+                except (OSError, IOError), e:
+                    _logger.error("%s: unable to create temporary stdout file: %s", self.label, e)
+                else:
+                    try:
+                        self.stdout_file = os.fdopen(stdout_fd, 'a+', 1)
+                    except (OSError, IOError), e:
+                        _logger.error("%s: unable to open temporary stdout file: %s", self.label, e)
+            finally:
+                if stdout_fn is not None:
+                    try:
+                        os.unlink(stdout_fn)
+                    except (OSError, IOError), e:
+                        _logger.warning("%s: unable to remove temporary stdout file: %s", self.label, e)
+            if self.stdout_file is None:
+                try:
+                    _logger.warning("%s: redirecting stdout to /dev/null", self.label)
+                    self.stdout_file = open("/dev/null")
+                except (OSError, IOError), e:
+                    _logger.error("%s: unable to open /dev/null (to redirect stdout): %s", self.label, e)
+                    raise
 
         stderr_fn = None
         try:
@@ -97,7 +113,7 @@ class SystemScriptForker (BaseForker):
     """
     Component for starting system script jobs such as the prologue and epilogue scripts run by cqm
     """
-    
+
     name = __name__.split('.')[-1]
     implementation = name
 
@@ -105,10 +121,10 @@ class SystemScriptForker (BaseForker):
 
     logger = _logger
 
-    def __init__ (self, *args, **kwargs):
+    def __init__(self, *args, **kwargs):
         """
         Initialize a new system script forker.
-        
+
         All arguments are passed to the base forker constructor.
         """
         BaseForker.__init__(self, *args, **kwargs)
